splitter ut beregn ytelse steg fra gammel stegservice
